Merge README.md into index.md and shorten
Both files contained mostly the same information. This commit merges those
into index.md, which is used by the website.

This commit shortens the 'wall of text'. By that, some details may be lost.
However, in the past years, not all rules were enforced, and this revision
does represent the practices from today and combines the ideas from the
original document into a structure that is more concise.

The idea is to not write down every detail, but to provide a
"rule of thumb" - a guideline - for minetest-mods.